CODAP-1260: split axis.spec.ts to reduce CI bottleneck
axis.spec.ts was the single-spec bottleneck on the regression CI workers
(9:34 vs ~7:30 for the other workers, on its own dedicated worker), and
this PR's new OTHER-bucket test made it longer. Split into three files so
Cypress can distribute them across parallel workers:

- axis.spec.ts (9 tests): single-attribute and paired-attribute add/remove
  with undo/redo (the simpler "Test graph axes with various attribute types"
  group)
- axis-splits.spec.ts (5 tests): split axes, multi-y attributes, bars/points
  domain adjustment, and OTHER-bucket overflow (the heavier configuration
  tests + the new CODAP-1260 test)
- axis-menu.spec.ts (9 tests): attribute-menu open/close/treat-as
  interactions + date-axis multi-y test

axis-menu.spec.ts (extracted from contexts 2 and 3) had a hidden
dependency on context 1's beforeEach to populate arrayOfAttributes/
arrayOfValues; the new file loads the fixture in its own beforeEach so it
stands alone.
